dropup working well now
