+++ components/DotMenu.vue
@@ -12,12 +12,13 @@ const props = defineProps({
     required: true,
   },
 })
+const emit = defineEmits(['update:data', 'changeEmit'])
 
 const dropupMenuRef = ref(null)
 </script>
 <template>
   <div class="card flex justify-content-center">
-    <Button
+    <!-- <Button
       class="text-cyan-500 hover:bg-cyan-50 relative"
       icon="pi pi-ellipsis-v"
       text
@@ -28,8 +29,8 @@ const dropupMenuRef = ref(null)
       @click="dropupMenuRef.openMenu()"
       aria-haspopup="true"
       aria-controls="overlay_menu"
-    />
-    <DropupMenu
+    /> -->
+    <!-- <DropupMenu
       ref="dropupMenuRef"
       :menuItems="props.menuItems"
       :label="props.label"
@@ -37,6 +38,14 @@ const dropupMenuRef = ref(null)
       <template #end>
         <slot name="end"></slot>
       </template>
-    </DropupMenu>
+    </DropupMenu> -->
+    <DropupMenuNew
+      v-model:data="props.menuItems"
+      :options="props.menuItems"
+      optionLabel="label"
+      :label="props.label"
+      @change="emit('changeEmit', $event)"
+      menu
+    />
   </div>
 </template>
